requirements.txt
anyio==4.4.0
APScheduler==3.10.4
certifi==2024.6.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
psycopg2-binary==2.9.9
python-telegram-bot==21.3
pytz==2024.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
typing_extensions==4.12.2
tzlocal==5.2
